% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_creators}
\alias{create_creators}
\title{create creators}
\usage{
create_creators(start)
}
\arguments{
\item{start}{The starting character. Essential in defining the type of
object the function should create.}
}
\description{
Function factory for creating fill string
}
\keyword{internal}
