% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_fill}
\alias{create_fill}
\title{create the final fill string from title, id, classes and attributes}
\usage{
create_fill(title = "title", id = "fj3", classes, attributes,
  json_ld = FALSE, function_container)
}
\arguments{
\item{title}{The title.}

\item{id}{The id.}

\item{classes}{Character vector with classes.}

\item{attributes}{Character vector with attributes.}

\item{function_container}{The list that contains the required functions
get_*}
}
\description{
create the final fill string from title, id, classes and attributes
}
\details{
The function_container is created with \code{create_creators}.
}
\keyword{internal}
