% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encl}
\alias{encl}
\title{enclose text by characters}
\usage{
encl(..., left = "{", right = "}")
}
\arguments{
\item{...}{The objects to be enclosed.}

\item{left}{The lhs-closing character.}

\item{right}{The rhs-closing character.}
}
\description{
enclose text by characters
}
\keyword{internal}
