% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{find_title}
\alias{find_title}
\title{elicit break titles via shiny gadget}
\usage{
find_title(level)
}
\arguments{
\item{level}{The level of the code break to be inserted}
}
\description{
A helper function to create a pane to enter a title name
}
