% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{give_breakchar}
\alias{give_breakchar}
\title{find breakchar for level}
\usage{
give_breakchar(level)
}
\arguments{
\item{level}{the level for which the break character should be returned}
}
\description{
minimal helper to return breakf for a given level
}
