% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{help_create_title}
\alias{help_create_title}
\title{create a title sequence}
\usage{
help_create_title(start = "##", fill = "this is a title",
  length = options()$strcode$char_length, sep = "sep_here", end = "----")
}
\arguments{
\item{start}{A sequence of letters to start the sequence}

\item{fill}{A sequence to fill the sequence with. This is the actual title.}

\item{length}{An integer value indicating how long the sequence should be}

\item{sep}{A separator sequence to separate start and fill.}

\item{end}{A character sequence that indicates how the end of the final
sequence should look like.}
}
\description{
This function returns a string that can be used as a title.
}
\details{
For Rstudio to recognize a herby produced sequence as a title, it must
    start with # and end with at least 4 of the following characters: #, -, =.
}
