% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{help_insert}
\alias{help_insert}
\title{help insert}
\usage{
help_insert(x, start_row = 1, start_indention = Inf,
  start_indention_margin = 0, end_row = 2, end_indention = Inf)
}
\arguments{
\item{x}{An object to insert}

\item{start_row}{the start row of the insertion}

\item{start_indention}{The start position within the row}

\item{start_indention_margin}{A margin (i.e. spaces) that will be added
at the target row before \code{x} is inserted.}

\item{end_row}{the row where the cursor should be after the insertion}

\item{end_indention}{The end position within the row}
}
\description{
A helper function to insert text
}

