% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{insert_break}
\alias{insert_break}
\alias{insert_l1_break}
\alias{insert_l2_break}
\alias{insert_l3_break}
\title{insert segment, section or subsection break}
\usage{
insert_l1_break()

insert_l2_break()

insert_l3_break()
}
\description{
A function designed to use as an RStudio
  \href{https://rstudio.github.io/rstudioaddins/}{add-in} for structuring code.
  There are three levels granularity
  \itemize{
    \item segments (level III), which are high-level blocks that can be separated
      as follows \cr    ### ======================
    \item sections (level II), which are medium-level blocks that can be separated
      as follows \cr    ## -----------------------------------------
    \item subsections (level I), which are low-level blocks that can be separated
      as follows \cr # ...................................................}
  For optimal use, we recommend specifying shortcuts in the add-in settings.
}
\details{
The breaks characters (\code{=}, \code{-}, \code{.}) were choosen
  such that they reflect the level of granularity, namely \code{===} has a
  much higher visual density than
  \code{...} \cr
  We recommend to start off grouping code into level II blocks.
  The advantage is that in both directions of granularity,
  there is another layer (\code{===} and \code{...}) left.
  When the code base grows, there
  might be the need to extend in both directions.
}
\examples{
  ## ===============================================
  ## Section 1: load data
  # ................................................
  # Subsection a: load first data set

  # [your code here]

  # ................................................
  # Subsection b: load second data set

  # [your code here]

  ## ==============================================
}

