% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{papply}
\alias{papply}
\title{vapply and paste}
\usage{
papply(..., collapse = " ")
}
\arguments{
\item{...}{arguments to be passed to vapply}

\item{collapse}{how to collapse the content via paste}
}
\description{
vapply and paste
}
\keyword{internal}
