% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strcode.R
\docType{package}
\name{strcode}
\alias{strcode}
\alias{strcode-package}
\alias{strcode-package}
\title{Structure and abstract your code better}
\description{
The \pkg{strcode} (short for structuring code) package contains
  tools to organize and abstract your code better.
}
\seealso{
See \code{vignette("strcode")} for an overview of the package.
}
\author{
Lorenz Walthert \email{lorenz.walthert@icloud.com}, \cr
}
