% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_str.R
\name{sum_str}
\alias{sum_str}
\title{Summarize the code structure}
\usage{
sum_str(dir_in = ".", dir_out = dir_in, file_in = "example.R",
  file_in_extension = ".R", file_out = NULL, file_out_extension = "",
  width = 50, line_nr = TRUE, granularity = 3, lowest_sep = TRUE,
  title = TRUE, ...)
}
\arguments{
\item{dir_in}{A directory where the file(s) can be found.}

\item{dir_out}{A directory where to print the output to. "" implies the
console.}

\item{file_in}{The name of a file which should be summarized. \code{NULL} will
create a summary of all files in the directory.}

\item{file_in_extension}{If \code{file_in} is \code{NULL}, all files with the
\code{file_in_extension} are considered.}

\item{file_out}{A connection or character string naming the file to print to.
The argument is irrelevant if \code{output_dir} is set to "".}

\item{file_out_extension}{An file extension for the file to be created.}

\item{width}{The character width of the output.}

\item{line_nr}{A boolean value that indicates whether the line numbers should
be printed along with the structure summary.}

\item{granularity}{Indicating how many levels should be in the summary}

\item{lowest_sep}{A boolean value indicating whether the separating lines
should be reported along their comments or not.}

\item{title}{A boolean value indicating whether the reported summary should
contain a title or not.}

\item{...}{futher arguments to be passed from and to other methods, in
particular \code{\link{list.files}} for reading in multiple files.}
}
\description{
Create a summary of one or multiple code files by collapsing them into the
section separators and their comments.
}

