% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchoring.R
\name{get_anchor}
\alias{get_anchor}
\title{Creat anchor sequence}
\usage{
get_anchor(enclosing_start = "#< ", enclosing_end = " >#",
  range_input = .Machine$integer.max, nchar_output = NULL, ...)
}
\arguments{
\item{enclosing_start}{A string that specifies the left-hand side enclosing
of the hash}

\item{enclosing_end}{A string that specifies the right-hand-side enclosing of
the hash. If \code{NULL}, then the reverse of \code{enclosing_start} is used.}

\item{range_input}{A number giving the range from which to draw
a random number.}

\item{nchar_output}{A number indicating how long the hash sequene should be.
\code{NULL} indicates no limitation.}

\item{...}{further arguments to be passed to and from other methods, in
in particular to digest}
}
\description{
Helper function to create an anchor string
}
\keyword{internal}
