% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{help_create_break}
\alias{help_create_break}
\title{create a break sequence}
\usage{
help_create_break(start = "##", sep = " ", break_char = "-",
  length = options()$strcode$char_length, anchor_in_sep = FALSE)
}
\arguments{
\item{start}{A sequence of letters to start the sequence.}

\item{sep}{A separator sequence to separate start and break_char.}

\item{break_char}{A character (sequence) used to create the actual break.}

\item{length}{An integer value indicating how long the sequence should be.}

\item{anchor_in_sep}{whether or not a code anchor (that is, a hash) should
be inserted in the center of the separator.}
}
\description{
the idea of the helper function is to return a string of a line length that is
composed of the start character and the break_characters
}
\keyword{internal}
