% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_anchor.R
\name{insert_anchor}
\alias{insert_anchor}
\title{Insert a hash anchor}
\usage{
insert_anchor(length_random_input = .Machine$integer.max,
  enclosing_start = "#<", enclosing_end = ">#", ...)
}
\arguments{
\item{length_random_input}{A number giving the range from which to withdraw
a random number.}

\item{enclosing_start}{A string that specifies the left-hand side enclosing
of the hash}

\item{enclosing_end}{A string that specifies the right-hand-side enclosing of
the hash. If \code{NULL}, then the reverse of \code{enclosing_start} is used.}

\item{...}{further arguments to be passed to and from other methods, in
in particular to digest}
}
\description{
Insert a hash anchor
}
\details{
The function draws a random number using sample from the range of
  \code{length_random_input}. Then, this number is hashed by calling the
  \code{digest} hash-function from the \code{digest} package. This hash
  is then wrapped between \code{enclosing_start} and \code{enclosing_end}.
}
\examples{
\dontrun{
  # RStudio needs to run for this
  insert_anchor(enclosing_start = "#!")
  }
}

