% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{insert_break}
\alias{insert_break}
\alias{insert_l1_break}
\alias{insert_l2_break}
\alias{insert_l3_break}
\title{insert segment, section or subsection break}
\usage{
insert_l1_break()

insert_l2_break()

insert_l3_break()
}
\description{
A function designed to use as an RStudio
\href{https://rstudio.github.io/rstudioaddins/}{add-in} for structuring code.
 \cr
There are three levels of granularity:
\itemize{
  \item level 1 sections, which are high-level blocks denoted by \cr
    ### ______________________
  \item level 2 sections, which are medium-level blocks denoted by \cr
    ##  ..............................................
  \item level 3 sections, which are low-level blocks denoted by \cr
    #   .. . . . . . . . . . . . . . . . . . . . . . . .
}
For optimal use, we recommend specifying keyboard shortcuts in the add-in
settings.
}
\details{
The breaks characters (\code{___}, \code{...}, \code{. .}) were
 chosen such that they reflect the level of granularity, namely \code{___}
 has a much higher visual density than \code{. .} \cr
 We recommend starting off by grouping code into level 2 blocks.
 The advantage is that in both directions of granularity, there is another
 layer (\code{___} and \code{...}) left. When the code base grows, there
 might be a need to extend in both directions.
}
\examples{
# This is a minimal example.
# See the readme for a longer and more detailed example.

##  ......................................................
##  A: pre-process t2
### .. . . . . . . . . . . . . . . . . . . . . . . . . . .
### a: substep 1



# [your code here]



### .. . . . . . . . . . . . . . . . . . . . . . . . . . .
### b: substep 2



# [your code here ]



##  ......................................................
}

