% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{insert_break}
\alias{insert_break}
\alias{insert_l1_break}
\alias{insert_l2_break}
\alias{insert_l3_break}
\title{insert segment, section or subsection break}
\usage{
insert_l1_break()

insert_l2_break()

insert_l3_break()
}
\description{
A function designed to use as an RStudio
  \href{https://rstudio.github.io/rstudioaddins/}{add-in} for structuring code.
  There are three levels granularity
  \itemize{
    \item level 1 sections, which are high-level blocks that can be separated
      as follows \cr    ### ______________________
    \item level 2 sections, which are medium-level blocks that can be separated
      as follows \cr    ##  ..............................................
    \item level 1 sections, which are low-level blocks that can be separated
      as follows \cr #   .. . . . . . . . . . . . . . . . . . . . . . . .}
  For optimal use, we recommend specifying shortcuts in the add-in settings.
}
\details{
The breaks characters (\code{___}, \code{...}, \code{. .}) were choosen
  such that they reflect the level of granularity, namely \code{___} has a
  much higher visual density than
  \code{. .} \cr
  We recommend to start off grouping code into level 2 blocks.
  The advantage is that in both directions of granularity,
  there is another layer (\code{===} and \code{...}) left.
  When the code base grows, there
  might be the need to extend in both directions.
}
\examples{
# this is a minimal example.
See the readme on \\href{https://github.com/lorenzwalthert/strcode}{gitub}
for a longer one.

##  ......................................................
##  A: pre-process t2
### .. . . . . . . . . . . . . . . . . . . . . . . . . . .
### a: substep 1



# [your code here]



### .. . . . . . . . . . . . . . . . . . . . . . . . . . .
### b: substep 2



# [your code here ]



##  ......................................................
}

