% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_breaks.R
\name{insert_break}
\alias{insert_break}
\title{Insert a code break of arbitrary level}
\usage{
insert_break(level)
}
\arguments{
\item{level}{The level, a numeric value bounded by 1 and 3}
}
\description{
A helper function to insert a code break for a given level
}
