% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_str.R
\name{sum_str_helper}
\alias{sum_str_helper}
\title{helper function for code summarisation
return code summary for one file}
\usage{
sum_str_helper(path_in, dir_out, file_out, file_out_extension, rm_rh_hashes,
  rm_rh_spaces, width, line_nr, granularity, last_sep, title, header)
}
\arguments{
\item{path_in}{Either a path to a directory or to a single file. If it is
a directory path, all files in that directory will be summarised. If it
is a single file path, only the resepective file will be summarised.
The default value uses the RStudio API to produce a summary of content from
the source editor. This requires that the file is saved before
\code{sum_str} is called.}

\item{dir_out}{The directory to print the output to. "" implies the console.}

\item{file_out}{A connection or character string naming the file to print to.
If set to \code{NULL}, the name will be \code{paste0("code-summary_", file_in)}.
If \code{dir_out} is set to "", \code{file_out} can be set to
"object" and the output of the function will be returned as an
object instead of just printed to the console with \code{cat}. This is
useful if output should be assigned to an object. If not set to "object",
\code{cat} will be used.}

\item{file_out_extension}{A file extension for the file to be created.}

\item{rm_rh_hashes}{Boolean value indicating whether or not to remove
righthand hashes in section titles (see details).}

\item{rm_rh_spaces}{Boolean value indicating whether or not to remove
righthand spaces in section titles (see details).}

\item{width}{The character width of the output. If NULL, it is set to the
length of the longest separator title.}

\item{line_nr}{A boolean value that indicates whether the line numbers should
be printed along with the structure summary.}

\item{granularity}{Indicates the lowest level of granularity that should be
included in the summary.}

\item{last_sep}{A boolean value indicating whether or not the separating
lines of the highest granularity should be printed.}

\item{title}{A boolean value indicating whether the reported summary should
contain a title or not.}

\item{header}{A boolean value indicating whether a column header should
indicate the name of the columns (line, level, section).}
}
\description{
Function is called by \code{sum_str()} and returns summary of one code file.
}
\details{
The core of the function is described best as follows: after a file
was read in and stored in a vector *lines* whereas each element describes a
line of code, the candidate lines (in the sense that they might be contained
in the summary) were evaluated, their indices stored in *cand*. Next,
various regex patterns are matched against *lines*. Then,
after all tests were executed, the variable *remove* contains all indices
that should be removed from *lines* before it is returned as the summary of
the code file. Hence, applying \code{setdiff(cand, remove)} contains
the subset of *lines* that we finally want to output.
}
\keyword{internal}
